[Refactor] effect-player-util.h 削除と整理。
