+++ src/effect/effect-player-oldies.cpp
@@ -1,5 +1,5 @@
 ﻿#include "effect/effect-player-oldies.h"
-#include "effect/effect-player-util.h"
+#include "effect/effect-player.h"
 #include "game-option/birth-options.h"
 #include "hpmp/hp-mp-processor.h"
 #include "monster-race/race-indice-types.h"
